Move PolygonTest from model.v1 package to projection.model package merge with existing
